COMMON: Move language GUIO code to language.h
