+++ common/language.cpp
@@ -20,6 +20,7 @@
  */
 
 #include "common/language.h"
+#include "common/gui_options.h"
 #include "common/str.h"
 
 namespace Common {
@@ -102,6 +103,23 @@ const char *getLanguageDescription(Language id) {
 			return l->description;
 	}
 	return 0;
+}
+
+bool checkGameGUIOptionLanguage(Language lang, const String &str) {
+	if (!str.contains("lang_")) // If no languages are specified
+		return true;
+
+	if (str.contains(getGameGUIOptionsDescriptionLanguage(lang)))
+		return true;
+
+	return false;
+}
+
+const String getGameGUIOptionsDescriptionLanguage(Language lang) {
+	if (lang == UNK_LANG)
+		return "";
+
+	return String("lang_") + getLanguageDescription(lang);
 }
 
 } // End of namespace Common
